Fixes NPE Crash on disable
This commit fixes a NullPointerException crash with DwarfCraft. This luckily only happens on plugin disable. This is caused by the onEnable method calling onDisable early before the trait information is available such as in the case when a required plugin such as Vault, PermsEx, or Citizens is not found. Since the trait information is not set yet the deregisterTrait() call in onDisable will crash with a NPE. Since the trait isn't registered until later in onEnable it is safe to skip the deregisterTrait call if the trait information is null. This prevents the plugin from crashing when disabling due to a required plugin not being found.