Adds LogBlock soft dependency
Added a soft dependency for LogBlock in the plugin.yml. This ensures that Spigot will load LogBlock before DwarfCraft in the case where LogBlock is on the server. This is not a required dependency and DwarfCraft will continue to work and load in the case that LogBlock is not on the server. If LogBlock is not loaded before DwarfCraft, DwarfCraft will not be able find the plugin in its onEnable. Without this there is no guarantee by Spigot that LogBlock will load before DwarfCraft.