Fix for DwarfCraft block drops being sporadic/unpredictable. This change causes the behavior and location of block drops to more closely match that of Vanilla. This is in place to help prevent blocks popping out into other blocks and shaking due to being spawned inside another block or popping up onto other blocks.
This issue was very common when cutting wood for trees as the wood
would be spawned naturally with a random offset which would cause it to
fly on top of the tree instead of falling to the player. This issue also
occurred when mining as the block drops would shake inside nearby blocks.

Having the blocks spawn without offsets helps prevent a lot of
confusion as to where the block you mined went and since any new drops
will clump with stacks already in the world you would not see a lot of
the drops.