Improved console logging
getLogger() is no longer recommended for use by plugins and the ConsoleSender should be used instead. Added a new Util method that uses the console sender and switched all plugin output to the new system.

Messages have also been slightly tweaked for clarity and to help server admins understand whats happening at a quick glance.

Severe errors which affect the functionality of the plugin are now in red color to grab attention. Warnings which are potential issues are now in a gold color. Debug messages enabled with /dc debug are in purple and all other messages such as info messages are in white. Important success messages are colored in green.